requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
pydantic==2.1.1
python-dotenv==1.0.0
motor==3.1.1
certifi==2023.7.22
typing-extensions==4.7.1